build: Add merge SHA into build artifact
When building artifact, by default the name is always build-version.
This is a problem if you're working on multiple PRs for which
you download the artifact: It's now harder to know who is who.

This fixes it by adding the github ref (branchname) and
SHA of the virtual merge commit into the filename of the artifact.