build: Add logo and favicon
Also, add a separate README in the docs/assets directory to make clear
that the logos are trademarks or registered trademarks (this README
must be a plain-text file, rather than Markdown, to keep MkDocs from
rendering it as part of the generated documentation). Update the
contribution guide to the same effect.