fix: Change gui name in config file
The name for gui in settings did not match the one used in the panel. This caused issues in the documentation, as the title differed from the one seen in cleura.cloud, which was confusing. The name was changed to now match the one seen in the panel by users.