feat: Enable glightbox plugin
Without this plugin, MkDocs Material scales down images to match the
flow text width. With screenshots, that makes it difficult to read
text unless the reader right-clicks on the image and selects "open
image in new tab", which is a bit tedious and makes for poor UX.

Instead, enable glightbox, which allows us to easily zoom the image to
its full size, with one click. In addition, this enables us to cycle
through the images in a page, which is helpful if it contains
sequential screen shots.

Reference: https://blueswen.github.io/mkdocs-glightbox/