build: Use environment variable for toggling htmlproofer
The htmlproofer plugin increases documentation build times nearly
tenfold. This may be undesirable when making quick changes while
running "tox -e serve". Thus, introduce an environment variable,
DOCS_ENABLE_HTMLPROOFER, and disable the plugin if it is set to
false.

Since this means that mkdocs.yml no longer contains "true" or "false"
for a boolean variable, disable the yamllint "truthy" check for that
line.

Also, update the contribution guide to that effect.

References:
https://www.mkdocs.org/user-guide/configuration/#environment-variables
https://yamllint.readthedocs.io/en/stable/disable_with_comments.html#disabling-checks-for-a-specific-line