chore: Temporarily disable htmlproofer in GitHub Actions workflows
We are currently getting false positives from the htmlproofer plugin
when it runs from GitHub Actions workflows, in that it reports links
as dead that are evidently alive. This is probably due to bandwidth
throttling or DDOS protection misfiring on the target sites.

This patch should be reverted as soon as the issues subside.