chore: Drop "extra." and "config.extra." prefixes from macros
Prior to this change, we referred to macros defined in the "extra"
dictionary of mkdocs.yml either with the "extra." or even the
"config.extra." prefix.

As the mkdocs-macros-plugin documentation explains, this is actually
not necessary: any variables defined in "extra" can be referred to
without such a prefix, and it is only when we want to access other
variables in "config" (such as "config.repo_url") that we need to
fully qualify them.

Since the documentation sources become much more concise that way, use
the abbreviated (unqualified) syntax.

Note: referring to the variables in this manner obviously means that
in the "extra" dictionary we cannot have any top-level keys named
"config", "page", "navigation", "environment", "plugin", "git", or
"files", but we don't at this time, and avoiding those top-level keys
should not be a big issue in the future.

Reference:
https://mkdocs-macros-plugin.readthedocs.io/en/latest/pages/#configuration-variables