build: Fail on undefined macros
Prior to this change, the macros plugin would render undefined macros
verbatim. That is, if a page's markdown contained the string
"{{ foobar }}" and there was no variable in mkdocs.yml named
extra.foobar, then the resulting HTML would also contain the string
"{{ foobar }}" (including the double braces).

Clearly that is not desirable; instead we just want the build to break
when the macros plugin finds an undefined macro.

To that end:

(1) Set the on_undefined configuration option to "strict". This, by
    itself, would only cause MkDocs to print an error message if it
    finds an undefined macro.

(2) Set the on_error_fail configuration option to true by
    default. This makes the MkDocs build actually fail when an
    undefined macro is found.

(3) Allow users to revert to the default, more lenient behavior
    described in (1), by setting the DOCS_FAIL_ON_MACRO_ERROR
    environment variable to false.

References:
https://mkdocs-macros-plugin.readthedocs.io/en/latest/troubleshooting/#what-happens-if-a-variable-is-undefined
https://mkdocs-macros-plugin.readthedocs.io/en/latest/advanced/#can-i-make-mkdocs-macros-build-process-to-fail-in-case-of-error-instead-of-displaying-the-error-on-the-page