fix: Adjust "Creating security groups" page after functionality changes
The how-to page for "Creating security groups" had section about removing
"default" security group ingress rules, that allowed all traffic. This was
recently changed in CCMP, and for new accounts the default ingress rules
are only allowing traffic from the same security group.

Now, the "remove security group rules" section does not recommend removing
the rules, but only suggests that it is something that you can do, if you
for any reason want to.

Screenshots and CLI output was adjusted as well.

The screenshot showing remaining rules, after removal of default ingress
rules, had wrong set of rules removed. This was changed, and the screenshot
filename was fixed.

To handle the case for accounts created prior the change, there
is a section explaining how to adjust the permissive ingress rules in the
old default policy group. It also explains who is possibly affected.