feat: Enable social cards
It sometimes comes in handy to share a bit of documentation on social
media (or chat), and the social cards plugin enables mkdocs-material
to generate some fancy cards for that purpose.

The cards themselves are not particularly useful without a brief
summary of the page being shared, and mkdocs-material neither has a
facility to auto-generate short page summaries, nor a provision to
mark up part of the page content as the summary.

However, we can apply a flipped approach where we

1. set the page description (which goes on the social card) in the YAML
   front matter for a page,
2. Use the macros plugin to then incorporate the description into the
page body.

This is something we can obviously gradually add over time, so apply
this approach to just a handful of pages for the time being.

One caveat does apply here, which is that we cannot use macro
expansion *within* the YAML front matter, so we cannot use things like
the {{brand}} placeholder in a description.

References:
https://squidfunk.github.io/mkdocs-material/setup/setting-up-social-cards/
https://squidfunk.github.io/mkdocs-material/reference/#setting-the-page-description