fix: Restore edit button
Our contribution guide mentions an edit button on every page. However,
as of mkdocs-material 9, this edit button is no longer on by default
if repo_url is set, but it is toggled by a feature flag. It also uses
a different icon.

Enable the feature flag, and also update the contribution guide to
refer to the updated icon. In doing so, use a bit of Jinja trickery
with mkdocs-macros so that the in-text edit icon auto-updates with the
one in the configuration (in case we ever change it).

Reference:
https://squidfunk.github.io/mkdocs-material/setup/adding-a-git-repository/#code-actions