feat: Add links to OpenStack API reference docs
Since it looks a bit odd for the Cleura Cloud API reference to freely
float in the reference section, also add a link to the OpenStack API
reference.

In doing so, create a new sub-section "api" under "reference", and
change the URL path for the Cleura Cloud API reference from
"/reference/cc-rest-api" to "/reference/api/cc".