build: Remove PDF export plugin
We've had PDF export disabled in our builds for a while now, and
nobody appears to be missing it.

Thus, remove the PDF export plugin, which also means we no longer
need to install the weasyprint package in the build environment
(which, in turn, speeds up the CI build checks).