build: Disable the no-duplicate-heading Markdown lint rule
The rationale for this rule is this:

"While not common to the default installation of most parsers, a
common extension to those parsers is to generate an id attribute for
the heading element and/or an anchor tag based on the content of the
heading element. Therefore, having two or more headings with the same
text may confuse those generators, producing poor results for those
anchors."

This does not apply to MkDocs and mkdocs-material, where if a single
page contains two headings like "Foo Bar", the corresponding anchors
will be named "#foo-bar" and "#foo-bar_1".

Thus, this rule should be safe to disable.

Reference:
https://github.com/jackdewinter/pymarkdown/blob/main/docs/rules/rule_md024.md