fix: Update information on object lock availability
Update the feature reference, reflecting the fact that object lock
is now available in Compliant Cloud regions.

Since this means that object lock is now available in all regions
where we have object storage at all, it is also safe to remove the
"available in select regions" caveat from the object lock how-to
guide.