build: Re-enable htmlproofer plugin in GitHub Actions workflow
Some testing with GitHub Actions seems to indicate that the issue
might not be affecting us anymore.

It is also possible that this is an issue in the plugin itself:
https://github.com/manuzhang/mkdocs-htmlproofer-plugin/issues/63

At any rate, we can try re-enabling the link checker and see if the
issue comes back in the future.

This reverts commit b87ec4e9620ac6e8549a0d619d34fed1920aad2d.