feat: Add tutorial on managing cloud objects with Ansible
We introduce a tutorial on Ansible, guiding the reader through creating
and managing various Cleura Cloud objects like keypairs, networks, and
servers. We also instruct the user to update a server, automatically
reboot if necessary, and install a Systemd service. Finally, the user is
instructed to delete Cleura Cloud objects --- always with the help of
Ansible.