build: Drop envdir from testenv
In tox<4, it was possible to share an single envdir for multiple
testenvs. Since tox 4, that ability has been removed, and if the same
envdir is used by multiple runs in different testenvs, the whole
deps list is reinstalled anyway.

Thus, it no longer makes sense to use a shared envdir. Drop it.