build: Fix thinko, don't create CSV export from GitHub Actions
Commit 911d9955de28690407f13ae0d66d047c19da2287 was misguided in the
sense that `continue-on-error: true` of course causes the report file
to be generated, but it also doesn't fail the build — and thus we
wouldn't be getting a notification if we had dead links on the site.

That's clearly not intentional, so drop CSV file generation from
`.linkcheckerrc`, and also remove the artifact upload step from the
`check` job. Instead, rely on the GitHub Actions job log to discern
which links are dead.

Getting a CSV file with a full report is still an option, when
running from a local checkout:

    tox -e check -- -F csv

Substitute other `-F` options like `html` or `xml` as needed.

References:
https://linkchecker.github.io/linkchecker/man/linkchecker.html#cmdoption-F
https://linkchecker.github.io/linkchecker/man/linkchecker.html#output-types