build: Rewrite link checker, make it run on commit
Prior to this change, the dead links check with linkchecker could only
be run post-merge. With this change, it checks for dead links on every
push, as part of the normal CI testing via GitHub Actions.

* Create a wrapper script around `linkchecker` named `linkcheck.sh`,
  and make it behave differently based on whether it's invoked as
  `linkcheck-local.sh` or `linkcheck-production.sh`.

* For `linkcheck-local.sh`, fire up a server on localhost:8000, and run
  `linkchecker` against that. Use that as part of the GitHub Actions
  `build` job, which runs on every push and PR.

* For `linkcheck-production.sh`, run linkchecker against
  https://docs.cleura.cloud. Invoke that from the `check` job, which
  runs on a schedule and can also be invoked manually.

Wrapping everything in a script also means we no longer need a
configuration file for `linkchecker`, so remove `.linkcheckerrc` from
the repository.