build: Post pull request comments on build check failure
Without this change, when a PR results in a build failure, the onus is
on the PR submitter to find the failed build, and the corresponding
build log. That is somewhat tedious and not particularly intuituive.

Instead, use a separate GitHub Action that posts a comment back to the
PR thread, containing a direct link to the failed build log. This
should hopefully make it easier for PR submitters to find out what
exactly their submitted change breaks.

References:
https://github.com/quipper/comment-failure-action/
https://github.com/quipper/comment-failure-action/pull/6