fix: Insert missing article in heading
Most how-to guides in the openstack/nova section use a heading along
the lines of "Doing X to a server", rather than "Doing X to server".

The how-to guide on restoring from a snapshot was the only deviation
from this. Inject the indefinite article into the heading, in order to
be in line with the other headings.