fix: Revert vGPU support to "in deployment, not yet supported"
Given the fact that it is not yet possible to launch a vGPU-enabled
server using CCMP, err on the side of caution and mark it as "in
deployment, not yet supported" rather than "in production and fully
supported".

However, now that we have a how-to guide for launching vGPU enabled
servers, link that how-to rather than just the flavor reference. Also,
update the how-to explaining that vGPU support is only available in
select regions, and add a backlink to the feature support matrix.