build: Remove the test-theme testenv
Using {posargs} in a deps variable breaks in the tox version that
ships on Ubuntu Focal. It does work on later versions, but we are only
using it in the test-theme testenv which we really don't need
anymore. Thus, by dropping that testenv we can remove this breakage,
and absolve people of having to pip-install tox on that platform.