build: Add validation parameters to mkdocs.yml
Since MkDocs 1.5, we can add a "validation" map to the configuration
file, which we can use to upgrade validation INFO messages to the
WARNING level — and this, in combination with the --strict setting in
the build testenv, can flag such issues as errors. This is what we
want, thus, set the values that MkDocs lists as "recommended settings
for most sites (maximal strictness)".

Reference:
https://www.mkdocs.org/user-guide/configuration/#validation