fix: Mention limit of 25 concurrently attached volumes
When launching a Nova server in Cleura Cloud, using one of the images
marked as public in Glance, the guest operating system gets VirtIO
block (virtio-blk) devices to work with. In the guest operating
system, these images are named /dev/vda for the boot block device,
and /dev/vdb, /dev/vdc etc. for attached volumes. Since VirtIO only
enumerates devices up to /dev/vdz, this means it can only handle 26
devices in total, and thus a single server can only ever attach 25
volumes concurrently.

Until we have the option of running from public images configured with
virtio-scsi (which does not have this limitation), state that 25 is
the maximum umber of volumes that can be concurrently attached to a
single server.