feat: Add redirects
We should have the ability to change the names of Markdown sources
and, hence, the URLs of rendered pages. However, in doing so we must
make sure that existing links to our documentation do not break.

The MkDocs "redirects" plugin enables exactly that. Add it to the
dependencies in `tox.ini`, enable it in `mkdocs.yml`, and also add a
note to the contribution guide reminding explaining how to add a
redirect.