refactor: Rename change-billing-data.md to change-account-data.md
The heading (and navigation entry) for this page has been "Changing
your account data" for a while, it makes sense for the URL to say the
same. Thus, rename the Markdown source, changing the URL.

Also, update internal links pointing to that page.
As for external links, add a redirect from the old to the new path.