build: Make linkchecker handle HTTP redirects gracefully
Since linkchecker 10.3.0, external links that are being redirected
(that is, they result in HTTP 301 or 302, rather than 200) result in a
warning. Since we are running linkchecker in strict mode (that is, not
ignoring warnings), this means the link check now fails on such links.

This is not good, however it would also not be good to just ignore all
warnings, as we would if we ran with --no-warnings. We just don't want
to fail on redirects.

Unfortunately, the `ignorewarnings` option in the linkchecker
configuration file, which allows us to ignore only *specific* warnings
(like `http-redirected`, in this case) has no equivalent CLI option.

Thus, we must bring back the .linkcheckerrc file (which we dropped in
49eeb6ef8d16b77e28751bcb905b6f053584f580), and set that option in the
config file.

References:
https://linkchecker.github.io/linkchecker/man/linkchecker.html#return-value
https://linkchecker.github.io/linkchecker/man/linkcheckerrc.html#filtering
https://github.com/linkchecker/linkchecker/pull/750