chore: Use material.extensions.emoji
Up until mkdocs-material 9.3, "materialx.emoji" was the correct way to
refer to the emoji generator. As of 9.4, we must should use
"material.extensions.emoji", as the old reference is deprecated.

Fix mkdocs.yml accordingly. Also, update tox.ini to make sure that
local checkouts are updated with an mkdocs-material version that
understands the new reference.