fix: Update the config-drive guide in light of CCMP changes
We could specify the config-drive configuration method for
new servers *only* when working with the OpenStack CLI.
Since there is now a way to do the exact same thing while
using the Cleura Cloud Management Panel, we update this
How-To guide accordingly.

While at it, we modify the server creation How-To guide to
mention the new config-drive configuration method and link
to this guide for more information.