fix: Drop :8080 port suffix from S3 how-to guides
Now that all our object storage services are listening on the regular
HTTPS port (443), we no longer need to list the :8080 suffix in our
how-to guides for S3.

We still do need to keep the suffix in the how-to guides for Swift,
since our object-storage endpoints in our service catalogs still do
contain the :8080 suffix.