fix: Update the object locking "legal hold" references
We previously referred the use of "put-legal-hold" aws cli command
which is wrong. The correct command to use is "put-object-legal-hold"
since you can only apply a legal hold to a single object at a time.
Also included an alternative way to set the legal hold for multiple
objects using the aws cli.