fix: Refresh invoice management How-To guide
We now show how to change the email address of an account in the
guide regarding account data, so we remove the corresponding
section from the invoice management guide. While at it, we adopt
the one sentence per line policy, use Font Awesome codes for
referring to UI icons, and introduce minor changes in the prose.