Update volume type reference
Updated the volume type reference to more accurately represent the types exposed in the API. Specifically made sure that the separate volume types with (`-encrypted`) is represented in this table, and that the default type is not encrypted. I am assuming that since there are volume types specifically ending in `-encrypted` the ones without the suffix is unencrypted.