build: Update GitLab CI for mike
Update the build workflow for mike.

This requires that we publish to GitLab pages from the gh-pages
branch, and updating that branch from a CI job in turn depends on a
GitLab project setting and the allow_push_repository_for_job_token
feature flag that, at least as of GitLab 17.x, is off by default.

References:
https://docs.gitlab.com/ci/jobs/ci_job_token/#allow-git-push-requests-to-your-project-repository
https://gitlab.com/gitlab-org/gitlab/-/issues/389060