build: Allow long body lines in commit messages
Since we're doing one sentence per line in Markdown sources, it makes no sense to not also allow that in commit messages.

Thus, disable (ignore) the gitlint body-max-line-length rule that enforces a maximum line length for the commit message body.

Retaining the corresponding rule for commit message subjects (title-max-length) is still a good idea because it promotes brevity in those.