fix: Use console blocks consistently
There are several instances where we use the bash/plain code block
combination to represent a command with its corresponding terminal
output. Instead, we use the console code block.

While at it, we fix a bug where backslash characters erroneously
follow the semicolon of a for-loop, and hence are visible in the
rendered code block.