build: Introduce api_domain and api_region variables
For per-region documentation, it makes sense to be able to construct
the full API URL.

Thus, introduce the api_domain and api_region variables. Set
api_region for this branch to the string FIXME, indicating that it
must be overridden in each per-region branch.

Note that api_region is meant to be the region name as defined in the
Keystone catalog, which is normally title-cased (such as "Kna1"). If
the variable needs to be lowercased in a macro, "{{api_region|lower}}"
(rendered as "kna1") would be the correct way to use it.