fix: Mention setting the digest algorithm for Swift TempURLs
The object storage infrastructure in Cleura Cloud currently exposes
the Swift API without support for the SHA256 digest algorithm that is
the default used by recent versions of the CLI.

Thus, in order for Swift TempURLs to work, we must specify the SHA1
digest alorithm, by using the "--digest sha1" option with the
"swift tempurl" command.

This change should be reverted once all object storage servers in
Cleura Cloud support SHA256 for TempURLs.