fix: Fix incorrectly rendered list in Swift TempURL howto
Without this change, a nested list in the Swift TempURL how-to guided
is rendered as a flat list. To fix this, indent the list by 4 spaces.

This is a known issue in MkDocs, though not a bug: rather, it is due
to a deliberate design decision in python-markdown.

In other words, it is working as indented.

References:
https://github.com/mkdocs/mkdocs/issues/545
https://github.com/Python-Markdown/markdown/issues/3