build: Use a non-existent reference for the "default" version
We want to ensure that the warning banner shown unconditionally: it
either tells readers what region they are reading documentation for,
or alerts them to the fact that they are reading the "generic" docs
and should perhaps consider looking into region-specific details as
well.

The way to do that is to set the "default" version in mkdocs.yml to
one that doesn't actually exist, making *all* published versions
"outdated" per the versioning logic and thus forcing the version
warning.