fix: Drop version paths from OpenStack endpoint URLs
Although it is correct (per the OpenStack catalog) to include the URL
path's "/v<version>" prefix in the endpoint definition, the OpenStack
APIs return HTTP 401 "Unauthorized" when the prefix is used. This
breaks our link checker.

Not using the prefix instead yields a HTTP 200, listing the available
version endpoints, and doesn't break the link checker.

The one exception from this is the radosgw Swift API, which will
return HTTP 405 "Method Not Allowed" even when accessing the bare
URI. This we have to specifically exclude from the link checks.