fix: Modify references to non-existent volume types
We no longer support the "default" and "ceph_hdd" Cinder volume types.
Thus, we modify references to the aforementioned with references to
generic or supported volume types.

While at it, we point out that using retyping to convert an encrypted
volume to an unencrypted one, or vice versa, is not possible. Instead,
we suggest a simple two-step procedure for (manually) copying data
between an encrypted volume and an unencrypted one.