build: Add a script to check deviations between branches
Now that we have multiple target branches to publish, we will
occasionally need to check on commits that exist in one branch but not
in another (in Git parlance, "cherries"). This is primarily so we can
check whether any merges from main into a per-region branch are still
pending.

To that end, add a convenience script that can list all cherries by
which the other branches differ from the current one.