feat: Add guide regarding HA for shoot control plane, worker groups
For existing shoot clusters, we have the option to enable high
availability (HA) for the control plane. We add a guide describing
how exactly we do that, noting that once HA is enabled, it cannot
be disabled. In the same guide, we show how to add HA characteristics
to worker groups.