build: Exclude OpenStack docs from link checks
It's rather silly to have to do this, but if 90% of our build runs
fail on unexplained connection resets from docs.openstack.org, which
then resolve themselves for no apparent reasons, it doesn't help much
to keep those links included in link checks.

Thus, add docs.openstack.org to the ignore list for the time being.